Fixed issue 42. Renamed packages
